tests/wpt/web-platform-tests/css/css-flexbox/align-items-baseline-vert-rl-column-horz-items.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#baseline-rules">
<link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#cross-alignment">
<link rel="author" href="mailto:sammy.gill@apple.com" title="Sammy Gill">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<style>
#target {
  display: flex;
  flex-direction: column;
  width: 200px;
  writing-mode: vertical-rl;
  align-items: baseline;
  border: 3px solid black;
  font-family: monospace;
  font-size: 10px;
  line-height: 10px;
}
#target > :nth-child(1) {
  background: hotpink;
  writing-mode: horizontal-tb;
  padding-left: 30px;
  margin-left: 10px;
}
#target > :nth-child(2) {
  background: papayawhip;
  writing-mode: horizontal-tb;
}
</style>
<body onload="checkLayout('#target > div')">
<div id="target">
  <div data-offset-y="11">two<br>lines</div>
  <div data-offset-y="11">hello</div>
</div>
</body>
